+++ Civi/ActionProvider/Provider.php
@@ -60,6 +60,8 @@ class Provider {
       AbstractAction::DATA_MANIPULATION_TAG));
 	  $this->addActionWithoutFiltering('SetValue', '\Civi\ActionProvider\Action\Generic\SetValue', E::ts('Set Value'), array(
 	    AbstractAction::DATA_MANIPULATION_TAG));
+    $this->addActionWithoutFiltering('SetParameterValue', '\Civi\ActionProvider\Action\Generic\SetParameterValue', E::ts('Set Value from parameter'), array(
+      AbstractAction::DATA_MANIPULATION_TAG));
     $this->addActionWithoutFiltering('SetDateValue', '\Civi\ActionProvider\Action\Generic\SetDateValue', E::ts('Set date value'), array(
       AbstractAction::DATA_MANIPULATION_TAG));
     $this->addActionWithoutFiltering('ConcatDateTimeValue', '\Civi\ActionProvider\Action\Generic\ConcatDateTimeValue', E::ts('Concat (merge) a date and a time field to one field'), array(