+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Version 1.9 (not yet released)
 * Added condition Contact Has Tag.
 * Added contribution recur id parameter to create contribution action.
 * Added receive data parameter to create contribution action.
+* Added action to set value from a parameter.
 
 Version 1.8
 -----------